Add utilities to regularly publish stats around heap memory usage and the number of GC collections, supporting all the major GC algorithms

+++ src/main/java/com/dici/service/metrics/MetricReporter.java
@@ -8,7 +8,7 @@ import java.time.Duration;
 public interface MetricReporter {
     void beginReport();
 
-    void addCount(String metricName, int count);
+    void addCount(String metricName, long count);
 
     void addMetric(String metricName, double value);
